updates tokio-named-pipes dependency
updates the `tokio-named-pipes` dependency in the `cargo.toml` file of the `pallas-network` project.

the update allows for the replacement of the version with a git repository and branch. this allows for an easier development process as changes to `tokio-named-pipes` can be tested immediately without having to create and push a new version to a registry.

```diff
diff --git a/pallas-network/cargo.toml b/pallas-network/cargo.toml
index 174c2c8d3927..df048680edc9 100644
--- a/pallas-network/cargo.toml
+++ b/pallas-network/cargo.toml
@@ -24,7 +24,7 @@ tokio = { version = "1", features = ["net", "io-util", "time", "sync", "macros"]
 tracing = "0.1.37"

 [target.'cfg(windows)'.dependencies]
-tokio-named-pipes = "0.1.0"
+tokio-named-pipes = { git = "git@github.com:falcucci/tokio-named-pipes.git", branch = "main" }
 windows-sys = "0.48.0"

 [dev-dependencies]
```